Refactor major.minor.micro version parsing into a function
virParseVersionString uses virStrToLong_ui instead of sscanf.

This also fixes a bug in the UML driver, that always returned 0
as version number.

Introduce STRSKIP to check if a string has a certain prefix and
to skip this prefix.

+++ src/uml/uml_conf.h
@@ -44,7 +44,7 @@ struct uml_driver {
 
     int privileged;
 
-    unsigned int umlVersion;
+    unsigned long umlVersion;
     int nextvmid;
 
     virDomainObjList domains;
